requirements.txt
numpy~=1.25.2
pandas~=1.5.3
flask~=2.2.5
python-dotenv~=1.0.0: